services/voice-server/requirements.txt
pipecat-ai[google,groq,cartesia,elevenlabs,websocket,kokoro]==1.0.0
fastapi>=0.115,<1.0
uvicorn>=0.32,<1.0
python-dotenv>=1.0,<2.0
pytest>=8.0
pytest-asyncio>=0.24
httpx>=0.27
